+++ packages/picker/src/index.ts
@@ -3,6 +3,7 @@ export * from './utils/color.ts'
 export type * from './utils/types.ts'
 
 export { useVModel } from './composables/useVModel.ts'
+export { parseHex } from './utils/parsers.ts'
 export { createUiSlots } from './utils/styles.ts'
 export type { Styling, VanillaClassFn, UiSliderSlots, UiInputSlots } from './utils/styles.ts'
 export { SVG_MOSAIC_URL } from './utils/consts.ts'